[FE] [FIX] address input and range slider
