+++ packages/visitorWeb/src/pages/MapPage/MapPage.tsx
@@ -227,6 +227,7 @@ const MapPage = () => {
             filter={getFilter()}
             categories={categories}
             allergens={allergens}
+            onChangeUserPosition={null}
           />
         </div>
         <MapView data={filteredRestaurants} />
